ml_advanced_training/requirements.txt
pandas==2.1.4
numpy==1.26.4
pandas-ta==0.3.14b0
scikit-learn==1.4.1
requests==2.31.0
joblib==1.3.2
pyyaml==6.0.1
python-dotenv==1.0.1
